+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## neptune-tensorflow-keras 0.9.9 [UNRELEASED]
+
+### Fixes
+- Small fixes and updates to NeptuneCallback docstrings ([#6](https://github.com/neptune-ai/neptune-tensorflow-keras/pull/6))
+
 ## neptune-tensorflow-keras 0.9.8
 
 ### Features